My apologies.
In the classic land of "it works for me," I never bothered testing it
on a fresh install.  Turns out the user settings file I was using
managed to make it work properly, but on a fresh install, this file was
empty, and nothing worked simply out of the box.

I'm charmed that it looks like there're people who attempted to install
this thing, hopefully it'll work better in the future.

 * Added proper defaults into the plugin itself
 * Settings no longer flush if nothing's changed
 * Fixed the preferences menu title
 * Added some messages to the console to help diagnose simple problems
 * Skip files that're too big to parse (nobody wants to scan a 100MB
log file for XML syntax)